feat: update dependencies and improve project structure
- Updated dependencies in package.json for better compatibility and performance.
- Refactored project entry keys from slug to id for consistency across components.
- Introduced content configuration for work, projects, and legal collections.
- Removed deprecated middleware for security headers; added headers configuration in vercel.json.
- Enhanced article layouts to utilize new rendering methods.
- Updated sitemap and RSS feeds to reflect changes in project identifiers.
