+++ astro.config.mjs
@@ -1,54 +1,73 @@
 import mdx from "@astrojs/mdx";
 import sitemap from "@astrojs/sitemap";
-import tailwindcss from '@tailwindcss/vite';
-import vercel from '@astrojs/vercel';
-import react from '@astrojs/react';
+import tailwindcss from "@tailwindcss/vite";
+import vercel from "@astrojs/vercel";
+import react from "@astrojs/react";
 import { defineConfig } from "astro/config";
 
 // https://astro.build/config
 export default defineConfig({
-	site: "https://kurtcalacday.vercel.app",
+  site: "https://kurtcalacday.vercel.app",
 
-	integrations: [
-		mdx(),
-		react(),
-		sitemap(),
-	],
+  integrations: [mdx(), react(), sitemap()],
 
-	vite: {
-		plugins: [
-			tailwindcss(),
-		],
-		build: {
-			sourcemap: false,
-		},
-	},
+  vite: {
+    plugins: [tailwindcss()],
+    build: {
+      sourcemap: false,
+    },
+  },
 
-	prefetch: {
-		prefetchAll: true,
-	},
+  prefetch: {
+    prefetchAll: true,
+  },
 
-	// Security Headers
-	server: {
-		headers: {
-			"X-Content-Type-Options": "nosniff",
-			"X-Frame-Options": "SAMEORIGIN",
-			"X-XSS-Protection": "1; mode=block",
-			"Strict-Transport-Security":
-				"max-age=63072000; includeSubDomains; preload",
-			"Referrer-Policy": "strict-origin-when-cross-origin",
-			"Permissions-Policy": "geolocation=(), microphone=(), camera=()",
-			"Content-Security-Policy":
-				"default-src 'self'; script-src 'self' 'unsafe-inline' https://static.cloudflareinsights.com https://cdn.vercel-insights.com https://app.rybbit.io data:; style-src 'self' 'unsafe-inline'; connect-src 'self' https://cloudflareinsights.com https://vitals.vercel-insights.com https://static.cloudflareinsights.com https://app.rybbit.io https://cdn.vercel-insights.com; img-src 'self' data: https:; object-src 'none';",
-		},
-	},
+  // Security Headers
+  server: {
+    headers: {
+      "X-Content-Type-Options": "nosniff",
+      "X-Frame-Options": "SAMEORIGIN",
+      "X-XSS-Protection": "1; mode=block",
+      "Strict-Transport-Security":
+        "max-age=63072000; includeSubDomains; preload",
+      "Referrer-Policy": "strict-origin-when-cross-origin",
+      "Permissions-Policy": "geolocation=(), microphone=(), camera=()",
+    },
+  },
 
-	// Vercel Configuration
-	output: "server",
-	adapter: vercel({
-		imageService: true,
-		webAnalytics: {
-			enabled: true,
-		},
-	}),
+  security: {
+    csp: {
+      directives: [
+        "default-src 'self'",
+        "img-src 'self' data: https:",
+        "font-src 'self'",
+        "connect-src 'self' https://cloudflareinsights.com https://vitals.vercel-insights.com https://static.cloudflareinsights.com https://app.rybbit.io https://cdn.vercel-insights.com",
+        "frame-src 'none'",
+        "object-src 'none'",
+        "form-action 'self'",
+      ],
+      scriptDirective: {
+        resources: [
+          "'self'",
+          "'unsafe-inline'",
+          "https://static.cloudflareinsights.com",
+          "https://cdn.vercel-insights.com",
+          "https://app.rybbit.io",
+          "data:",
+        ],
+      },
+      styleDirective: {
+        resources: ["'self'", "'unsafe-inline'"],
+      },
+    },
+  },
+
+  // Vercel Configuration
+  output: "server",
+  adapter: vercel({
+    imageService: true,
+    webAnalytics: {
+      enabled: true,
+    },
+  }),
 });
