+++ package.json
@@ -1,39 +1,42 @@
 {
-	"name": "portfolio-website",
-	"type": "module",
-	"version": "1.0.0",
-	"scripts": {
-		"dev": "astro dev",
-		"build": "astro build",
-		"preview": "astro preview",
-		"astro": "astro"
-	},
-	"dependencies": {
-		"@astrojs/mdx": "4.3.13",
-		"@astrojs/react": "^4.4.2",
-		"@astrojs/rss": "4.0.15",
-		"@astrojs/sitemap": "3.7.0",
-		"@astrojs/vercel": "9.0.4",
-		"@base-ui/react": "^1.2.0",
-		"@tailwindcss/vite": "^4.2.1",
-		"@types/react": "^19.2.14",
-		"@types/react-dom": "^19.2.3",
-		"@vercel/analytics": "^1.6.1",
-		"astro": "5.18.0",
-		"class-variance-authority": "^0.7.1",
-		"clsx": "^2.1.1",
-		"cmdk": "^1.1.1",
-		"fuse.js": "^7.1.0",
-		"lucide-react": "^0.556.0",
-		"react": "^19.2.4",
-		"react-dom": "^19.2.4",
-		"simple-icons": "^16.11.0",
-		"tailwind-merge": "^3.5.0",
-		"tw-animate-css": "^1.4.0",
-		"typescript": "^5.9.3"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "2.3.8",
-		"@types/bun": "^1.3.10"
-	}
+  "name": "portfolio-website",
+  "type": "module",
+  "version": "1.0.0",
+  "scripts": {
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro"
+  },
+  "dependencies": {
+    "@astrojs/mdx": "^4.3.14",
+    "@astrojs/react": "^5.0.2",
+    "@astrojs/rss": "^4.0.18",
+    "@astrojs/sitemap": "^3.7.2",
+    "@astrojs/vercel": "10.0.3",
+    "@base-ui/react": "^1.3.0",
+    "@tailwindcss/vite": "^4.2.2",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "@vercel/analytics": "^1.6.1",
+    "astro": "^6.1.1",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.1.1",
+    "fuse.js": "^7.1.0",
+    "lucide-react": "^0.556.0",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "simple-icons": "^16.14.0",
+    "tailwind-merge": "^3.5.0",
+    "tw-animate-css": "^1.4.0",
+    "typescript": "^5.9.3"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "2.3.8",
+    "@types/bun": "^1.3.11"
+  },
+  "engines": {
+    "node": ">=22.0.0"
+  }
 }
