+++ src/components/ArrowCard.tsx
@@ -48,7 +48,7 @@ export default function ArrowCard({
 
 	return (
 		<a
-			href={`/projects/${entry.slug}`}
+			href={`/projects/${entry.id}`}
 			className="group p-4 gap-3 flex items-center border rounded-lg hover:bg-black/5 hover:dark:bg-white/10 border-black/15 dark:border-white/20 transition-colors duration-300 ease-in-out"
 		>
 			<div className="w-full group-hover:text-black group-hover:dark:text-white blend">
