+++ src/components/Search.tsx
@@ -12,7 +12,7 @@ type Props = {
 };
 
 function getProjectEntryKey(entry: ProjectEntryWithPreview): string {
-	return entry.slug ?? entry.id ?? entry.data.title;
+	return entry.id ?? entry.data.title;
 }
 
 export default function Search({ data, tagOptionsByEntry }: Props) {
