+++ src/pages/legal/[...slug].astro
@@ -1,5 +1,5 @@
 ---
-import { type CollectionEntry, getCollection } from "astro:content";
+import { type CollectionEntry, getCollection, render } from "astro:content";
 import BottomLayout from "@layouts/BottomLayout.astro";
 import PageLayout from "@layouts/PageLayout.astro";
 import TopLayout from "@layouts/TopLayout.astro";
@@ -11,33 +11,33 @@ export const prerender = true;
 
 // Create the static pages for legal docs
 export async function getStaticPaths() {
-	const docs = await getCollection("legal");
-	return docs.map((doc: { slug: string }) => ({
-		params: { slug: doc.slug },
-		props: doc,
-	}));
+  const docs = await getCollection("legal");
+  return docs.map((doc: { id: string }) => ({
+    params: { slug: doc.id },
+    props: doc,
+  }));
 }
 
 // Get the requested legal doc
 type Props = CollectionEntry<"legal">;
 const doc = Astro.props;
 const { title, date } = doc.data;
-const { Content } = await doc.render();
+const { Content } = await render(doc);
 ---
 
 <PageLayout title={title} description={`${title} for ${SITE.TITLE}`}>
   <TopLayout>
-    <div class='animate'>
-      <div class='page-heading'>
+    <div class="animate">
+      <div class="page-heading">
         {title}
       </div>
-      <p class='font-normal opacity-75'>
+      <p class="font-normal opacity-75">
         Last updated: {formatDate(date)}
       </p>
     </div>
   </TopLayout>
   <BottomLayout>
-    <article class='animate'>
+    <article class="animate">
       <Content />
     </article>
   </BottomLayout>
